cloud-functions/functions/goodpractices_thumb/requirements.txt
google-cloud-vision==3.1.0
google-cloud-storage==2.0.0; python_version < '3.7'
google-cloud-storage==2.1.0; python_version > '3.6'
Wand==0.6.8